Add GbSourceViewMode
This is an object representing some kind of modal keyboard state
in the source view. It can be used to hang GtkBindingSet on the mode
via CSS rules.

Its kind of a hack which inherits from GtkWidget just to get at the
GtkBindings code, but is not actually used as a widget.