libide: add basic completion provider for C
Lots of stuff isn't wired up yet, but I don't want this code to land in
the bit bucket.

In particular, the display code for the completion window needs an
incredible amount of work. Some of it will likely require changes in
GSV so that we can get columns for appropriate layout.

Also, very little effort seems to be going into the result sorting, so
items do not get shown in a very helpful way.