save modified buffers upon context unloading
we came to the design decision a while ago we'd like to just save the
files. what i havent decided yet, is if we should save the unsaved
changes to drafts (and reload on next start) or to just the file
directly. for now, we will go to the backing file directly.