load workspace even if no files are loaded
Long term, we are going to have a project selection window. But until
then, this will get us through our first release.

If the current directory is not the home dir, we use it. If the current
directory is the home dir and ~/Projects exists, we use ~/Projects.
Otherwise, we use the home directory.