highlight: add very crude, not even vaguely optimize highlight
This does tons of things wrong that need to be changed, but i dont want
to lose a snapshot of something "working".

We should change this to use a callback design, so we can amortize the
overhead cost of walking through the buffer and fetching the translation
unit.

We should also probably drop the intermediate IdeHighlightKind and allow
the highlighter to set the style type directly. Otherwise, we might lock
ourselves into only being able to highlight things that are "generic"
amongst all the languages.

For example, we might want "c:function" or "cpp:method" instead of
"def:function".

Also, the index is currently oversized, (about 350k on my current
working set). We can still find some tricks to prune this down.

Anyway, step 1!