highlight: implement incremental, callback based, highlight engine
Now, the highlight engine will provide a callback to perform a style
update. That callback will return if the highlighter may continue or
if it has expired its quanta.

This helps reduce the overhead in highlighters that need to setup a bunch
of information to make forward progress in the buffer. It has the small
cost of a qdata lookup in the callback. (We could make this less by
enforcing the highlighter pass extra data back, but I don't like the API
much for that).

The index can now take arbitrary pointers as the values for words. The
clang index takes advantage of this to specify the direct tags it wants.

This will change in the future as we add more styling information to
builder.xml (and builder-dark.xml). Other style schemes will need to
do the same if they want extra-detailed coloring.

For example, we have plans for functions, macros, members, fields, and
other special types which may not map to the same generic coloring that
other languages will want to use.

Anyway, WIP