project-selector: add basic clone support with libgit2-glib
We still need a bunch of work on the callbacks, but as a proof of concept
this works. You need to select an empty directory, which I'm not happy
about. I'd like to take the repository name in the path and add that to
the suffix of the directory path.

We could also do something like:

  [ URL ]
  [ Name ]
  [ Directory ]

and then modify Directory filechooserbutton when Name changes. This would
allow for specific overrides.