project-selector: don't use modal or transient for for new-project dialog
Setting the type-hint to dialog is enough to get the effect we want, which
is the dialog to always stay above the workbench windows. Making them
modal or transient-for implies that 1) all other windows should be blocked
and 2) that the operation is relative to the workbench. Neither of these
is true.

We do, however, want to stay transient-for the select-project window when
the "New" button in the header bar is pressed.