icons: bundle gitg icon as our git vcs icon
We need to check with the gitg team that this is okay with them. We intend
to use more and more from gitg, so I think it would be good to recognize
their work with the iconography as well.

Since I know many of you read this, ping me with alternate suggestions if
you have them!